GBTG.sol
The Golden Bit Token is a digital token issued by the Fourth Art Foundation of the United States to promote the circulation of artworks worldwide. This token is a token dedicated to the circulation of Chinese painting by the Fund for the Greater China region. This token can be exchanged for the authenticity of domestic masters. Gold Bit for Growth is abbreviated as GBTG.